add(CarName) : CarName 도메인 분리
